% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gpcp.R
\name{runGPCP}
\alias{runGPCP}
\title{Genomic Prediction of Cross Performance
This function performs genomic prediction of cross performance using genotype and phenotype data.}
\usage{
runGPCP(
  phenotypeFile,
  genotypeFile = NA,
  genotypeData = NA,
  genotypes,
  traits,
  weights = NA,
  userSexes = "",
  userFixed = NA,
  userRandom = NA,
  Ploidy = NA,
  NCrosses = NA
)
}
\arguments{
\item{phenotypeFile}{A data frame containing phenotypic data, typically read from a CSV file.}

\item{genotypeFile}{Path to the genotypic data, either in VCF or HapMap format.}

\item{genotypeData}{A dataframe containing genotypic data if genotypeFile not provided}

\item{genotypes}{A character string representing the column name in the phenotype file for the genotype IDs.}

\item{traits}{A string of comma-separated trait names from the phenotype file.}

\item{weights}{A numeric vector specifying weights for the traits.}

\item{userSexes}{A string representing the column name corresponding to the individuals' sexes.}

\item{userFixed}{A string of comma-separated fixed effect variables.}

\item{userRandom}{A string of comma-separated random effect variables.}

\item{Ploidy}{An integer representing the ploidy level of the organism.}

\item{NCrosses}{An integer specifying the number of top crosses to output.}
}
\value{
A data frame containing predicted cross performance.
}
\description{
Genomic Prediction of Cross Performance
This function performs genomic prediction of cross performance using genotype and phenotype data.
}
\examples{
\donttest{
# Load phenotype data from CSV
# Diploid pipeline
phenotypeFile <- read.csv(system.file("extdata", "phenotypeFile.csv", package = "gpcp"))
genotypeFile <- system.file("extdata", "genotypeFile_Chr9and11.vcf", package = "gpcp")
finalcrosses <- runGPCP(
    phenotypeFile = phenotypeFile,
    genotypeFile = genotypeFile,
    genotypes = "Accession",
    traits = "YIELD,DMC",
    weights = c(3, 1),
    userFixed = "LOC,REP",
    Ploidy = 2,
    NCrosses = 150
)
message(finalcrosses)
 #PolyPLoid Pipeline
 # 1) load example data from the package
data(DT_polyploid, package = "sommer")
DT <- DT_polyploid
GT <- GT_polyploid
MP <- MP_polyploid
# 2) convert A/T/C/G strings to numeric codes
numo <- sommer::atcg1234(data = GT, ploidy = 4)

# 3) find the set of individuals common to genotypes and phenotypes
common <- intersect(DT$Name, rownames(numo$M))
marks  <- numo$M[common, , drop = FALSE]
pheno2 <- as.data.frame(DT[match(common, DT$Name), ])
# 4) call runGPCP with ploidy = 4
result4x <- suppressWarnings(
  runGPCP(
    phenotypeFile = pheno2,
    genotypeData   = marks,
    genotypes      = "Name",
    traits         = c("total_yield", "tuber_length"),
    weights        = c(3, 1),
    Ploidy         = 4,
    NCrosses       = 100
  )
)
}
}
